backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "faiss-cpu>=1.13.2",
    "fastapi>=0.135.2",
    "httpx>=0.28.1",
    "networkx>=3.6.1",
    "pydantic-settings>=2.13.1",
    "pygit2>=1.19.1",
    "rank-bm25>=0.2.2",
    "tree-sitter>=0.25.2",
    "tree-sitter-java>=0.23.5",
    "torch>=2.10.0",
    "transformers>=5.3.0",
    "peft>=0.18.1",
    "websockets>=16.0",
    "uvicorn[standard]>=0.42.0",
    "datasets>=4.8.4",
    "trl>=0.29.1",
]

[[tool.uv.index]]
name = "pytorch-cpu"
url = "https://download.pytorch.org/whl/cpu"
explicit = true
